CoreComponents: Generalize node linkage in navigation maker classes. (Breaking change.)
Add BaseNavigationMarker and insert it as the base for navigation marker
components.  This generalizes the behavior so clients can use the
markers less knowledge of the implementation.  A client doesn't need to
know if the marker is a waypoint sequence of a node graph.

Breaking changes: The type of the serialized link fields for
NavigationWaypoint and NavigationNode have been changed to
BaseNavigationMarker.   This allows waypoint and node graphs to be
linked to each other.   This change will result in the loss of existing
links.