Core: Add ObjectList Feature
ObjectList: A list of Unity Objects that must be of a give type.  Useful
for serialization of an array of objects that are known to be Unity
Objects. but present a non-Unity Object interface during normal use.
E.g. A list of Unity Objects that all implement the IAgent interface.

ObjectListDrawer: A GUI element representing a reorderable list of
single-line object references.  Provides a user friendly GUI element for
ObjectList implementations.