Fixed:  'GravityBody' status body not taken into account.
'GravityBody' is when the collider is disabled, the rigidbody is
non-kinematic, and gravity is turned on.  This was showing up as
disabled, and SetStatus(Disabled) was not taking the gravity flag into
account.

Fixed the issue by adding the GravityBody type and adjusting the get/set
status methods to properly handle the new type.  Also added a
SetStatus(status, useGravity) overload since useGravity is now in the
mix.